requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
paramiko
proxmoxer
requests
